requirements.txt
Flask==2.2.5
Flask-Compress==1.10.1
cachetools==4.2.2
requests==2.26.0
python-dotenv==0.19.0
Werkzeug==2.2.3
Flask-SQLAlchemy==3.1.1
Pillow==10.0.0
gunicorn==21.2.0